LPS-54543 rename
